+++ .github/ISSUE_TEMPLATE/이슈-템플릿.md
@@ -9,10 +9,9 @@ assignees: ''
 
 ## 내용
 
-
 ## 작업 항목
-- [ ] 작업 내용
-- [ ] 작업 내용
 
+- [ ] 작업 내용
+- [ ] 작업 내용
 
 ## 참고 (선택)
