+++ .github/PULL_REQUEST_TEMPLATE.md
@@ -1,10 +1,11 @@
 ## 관련 이슈
+
 - closes #이슈번호
 
 ## 변경사항
 
-
 ## 테스트
+
 - [ ] 로컬 테스트 완료
 - [ ] API 문서(Swagger) 확인
