+++ build.gradle
@@ -49,7 +49,7 @@ dependencies {
 
     // Mail
     implementation 'org.springframework.boot:spring-boot-starter-mail'
-    
+
     //JSONB
     implementation 'io.hypersistence:hypersistence-utils-hibernate-63:3.7.0'
 }
