+++ docker-compose.yml
@@ -32,7 +32,7 @@ services:
     ports:
       - "5432:5432"
     healthcheck:
-      test: ["CMD-SHELL", "pg_isready -U hrbank"]
+      test: [ "CMD-SHELL", "pg_isready -U hrbank" ]
       interval: 5s
       timeout: 5s
       retries: 5
