+++ src/main/java/com/hrbank3/hrbank3/config/SchedulerConfig.java
@@ -6,4 +6,5 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 @Configuration
 @EnableScheduling
 public class SchedulerConfig {
+
 }
