+++ src/main/java/com/hrbank3/hrbank3/dto/dashboard/EmployeeDistributionDto.java
@@ -1,7 +1,9 @@
 package com.hrbank3.hrbank3.dto.dashboard;
 
 public record EmployeeDistributionDto(
-        String groupKey,
-        long count,
-        double percentage
-) {}
+    String groupKey,
+    long count,
+    double percentage
+) {
+
+}
