+++ src/main/java/com/hrbank3/hrbank3/dto/dashboard/EmployeeTrendDto.java
@@ -2,8 +2,10 @@ package com.hrbank3.hrbank3.dto.dashboard;
 
 // 직원 수 추이 응답 DTO
 public record EmployeeTrendDto(
-        String date,
-        long count,
-        long change,
-        double changeRate
-) {}
+    String date,
+    long count,
+    long change,
+    double changeRate
+) {
+
+}
