+++ src/main/java/com/hrbank3/hrbank3/dto/file/FileDownloadDto.java
@@ -6,4 +6,6 @@ public record FileDownloadDto(
     Resource resource,
     String originalName,
     String contentType
-) {}
+) {
+
+}
