+++ src/main/java/com/hrbank3/hrbank3/entity/enums/TrendUnit.java
@@ -2,5 +2,5 @@ package com.hrbank3.hrbank3.entity.enums;
 
 // 직원 수 추이 조회 시간 단위
 public enum TrendUnit {
-    day, week, month, quarter, year
+  day, week, month, quarter, year
 }
