--- src/main/java/com/hrbank3/hrbank3/repository/DepartmentRepositoryCustom.java
@@ -1,16 +1,0 @@
-package com.hrbank3.hrbank3.repository;
-
-import com.hrbank3.hrbank3.dto.CursorPageResponseDto;
-import com.hrbank3.hrbank3.dto.department.DepartmentDto;
-
-public interface DepartmentRepositoryCustom {
-
-    CursorPageResponseDto<DepartmentDto> findAllWithCursor(
-            String nameOrDescription,
-            String sortField,
-            String sortDirection,
-            Long idAfter,
-            String cursor,  // 추가
-            int size
-    );
-}
