+++ src/main/java/com/hrbank3/hrbank3/repository/custom/BackupHistoryRepositoryCustom.java
@@ -1,10 +1,11 @@
 package com.hrbank3.hrbank3.repository.custom;
 
+import com.hrbank3.hrbank3.dto.backup_history.BackupHistorySearchCondition;
 import com.hrbank3.hrbank3.entity.BackupHistory;
-import com.hrbank3.hrbank3.repository.condition.BackupHistorySearchCondition;
 import java.util.List;
 
 public interface BackupHistoryRepositoryCustom {
+
   List<BackupHistory> findAllByCondition(BackupHistorySearchCondition condition);
 
   long countByCondition(BackupHistorySearchCondition condition);
