+++ src/main/java/com/hrbank3/hrbank3/repository/custom/DepartmentRepositoryImpl.java
@@ -1,4 +1,4 @@
-package com.hrbank3.hrbank3.repository;
+package com.hrbank3.hrbank3.repository.custom;
 
 import com.hrbank3.hrbank3.dto.CursorPageResponseDto;
 import com.hrbank3.hrbank3.dto.department.DepartmentDto;
@@ -8,11 +8,10 @@ import com.hrbank3.hrbank3.entity.QEmployee;
 import com.hrbank3.hrbank3.entity.enums.EmployeeStatus;
 import com.querydsl.core.BooleanBuilder;
 import com.querydsl.jpa.impl.JPAQueryFactory;
-import lombok.RequiredArgsConstructor;
-import org.springframework.stereotype.Repository;
-
 import java.util.List;
 import java.util.stream.Collectors;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Repository;
 
 @Repository
 @RequiredArgsConstructor
@@ -22,12 +21,12 @@ public class DepartmentRepositoryImpl implements DepartmentRepositoryCustom {
 
   @Override
   public CursorPageResponseDto<DepartmentDto> findAllWithCursor(
-          String nameOrDescription,
-          String sortField,
-          String sortDirection,
-          Long idAfter,
-          String cursor,
-          int size) {
+      String nameOrDescription,
+      String sortField,
+      String sortDirection,
+      Long idAfter,
+      String cursor,
+      int size) {
 
     QDepartment department = QDepartment.department;
     QEmployee employee = QEmployee.employee;
@@ -36,16 +35,16 @@ public class DepartmentRepositoryImpl implements DepartmentRepositoryCustom {
     // 키워드 필터링
     if (nameOrDescription != null && !nameOrDescription.isBlank()) {
       builder.and(
-              department.name.containsIgnoreCase(nameOrDescription)
-                      .or(department.description.containsIgnoreCase(nameOrDescription))
+          department.name.containsIgnoreCase(nameOrDescription)
+              .or(department.description.containsIgnoreCase(nameOrDescription))
       );
     }
 
     // 전체 개수 조회
     long totalElements = queryFactory
-            .selectFrom(department)
-            .where(builder)
-            .fetchCount();
+        .selectFrom(department)
+        .where(builder)
+        .fetchCount();
 
     // 커서 조건 추가 (cursor가 있으면 cursor 사용, 없으면 idAfter 사용)
     if (cursor != null && !cursor.isBlank()) {
@@ -61,17 +60,17 @@ public class DepartmentRepositoryImpl implements DepartmentRepositoryCustom {
 
     // 정렬 및 조회
     List<Department> results = queryFactory
-            .selectFrom(department)
-            .where(builder)
-            .orderBy("desc".equals(sortDirection)
-                    ? ("establishedDate".equals(sortField)
-                    ? department.establishedDate.desc()
-                    : department.name.desc())
-                    : ("establishedDate".equals(sortField)
-                    ? department.establishedDate.asc()
-                    : department.name.asc()))
-            .limit(size + 1)
-            .fetch();
+        .selectFrom(department)
+        .where(builder)
+        .orderBy("desc".equals(sortDirection)
+            ? ("establishedDate".equals(sortField)
+            ? department.establishedDate.desc()
+            : department.name.desc())
+            : ("establishedDate".equals(sortField)
+                ? department.establishedDate.asc()
+                : department.name.asc()))
+        .limit(size + 1)
+        .fetch();
 
     // 다음 페이지 여부 확인
     boolean hasNext = results.size() > size;
@@ -81,36 +80,36 @@ public class DepartmentRepositoryImpl implements DepartmentRepositoryCustom {
 
     // employeeCount 실제 조회
     List<DepartmentDto> content = results.stream()
-            .map(d -> {
-              Long count = queryFactory
-                      .select(employee.count())
-                      .from(employee)
-                      .where(employee.department.eq(d),
-                              employee.status.ne(EmployeeStatus.RESIGNED))
-                      .fetchOne();
-              return new DepartmentDto(
-                      d.getId(),
-                      d.getName(),
-                      d.getDescription(),
-                      d.getEstablishedDate(),
-                      d.getCreatedAt(),
-                      d.getUpdatedAt(),
-                      count != null ? count : 0L,
-                      d.getDepartmentMail()
-              );
-            })
-            .collect(Collectors.toList());
+        .map(d -> {
+          Long count = queryFactory
+              .select(employee.count())
+              .from(employee)
+              .where(employee.department.eq(d),
+                  employee.status.ne(EmployeeStatus.RESIGNED))
+              .fetchOne();
+          return new DepartmentDto(
+              d.getId(),
+              d.getName(),
+              d.getDescription(),
+              d.getEstablishedDate(),
+              d.getCreatedAt(),
+              d.getUpdatedAt(),
+              count != null ? count : 0L,
+              d.getDepartmentMail()
+          );
+        })
+        .collect(Collectors.toList());
 
     String nextCursor = hasNext ? String.valueOf(results.get(results.size() - 1).getId()) : null;
     Long nextIdAfter = hasNext ? results.get(results.size() - 1).getId() : null;
 
     return new CursorPageResponseDto<>(
-            content,
-            nextCursor,
-            nextIdAfter,
-            content.size(),
-            totalElements,
-            hasNext
+        content,
+        nextCursor,
+        nextIdAfter,
+        content.size(),
+        totalElements,
+        hasNext
     );
   }
 }
