+++ src/main/java/com/hrbank3/hrbank3/repository/custom/EmployeeRepositoryCustom.java
@@ -1,7 +1,7 @@
 package com.hrbank3.hrbank3.repository.custom;
 
+import com.hrbank3.hrbank3.dto.employee.EmployeeSearchCondition;
 import com.hrbank3.hrbank3.entity.Employee;
-import com.hrbank3.hrbank3.repository.condition.EmployeeSearchCondition;
 import java.util.List;
 
 public interface EmployeeRepositoryCustom {
