+++ src/main/java/com/hrbank3/hrbank3/repository/custom/EmployeeRepositoryImpl.java
@@ -1,9 +1,9 @@
 package com.hrbank3.hrbank3.repository.custom;
 
+import com.hrbank3.hrbank3.dto.employee.EmployeeSearchCondition;
 import com.hrbank3.hrbank3.entity.Employee;
-import com.hrbank3.hrbank3.entity.enums.EmployeeStatus;
 import com.hrbank3.hrbank3.entity.QEmployee;
-import com.hrbank3.hrbank3.repository.condition.EmployeeSearchCondition;
+import com.hrbank3.hrbank3.entity.enums.EmployeeStatus;
 import com.querydsl.core.types.OrderSpecifier;
 import com.querydsl.core.types.dsl.BooleanExpression;
 import com.querydsl.jpa.impl.JPAQueryFactory;
