+++ src/main/java/com/hrbank3/hrbank3/service/EmployeeAuditHistoryService.java
@@ -4,6 +4,7 @@ import com.hrbank3.hrbank3.common.util.IpExtractUtil;
 import com.hrbank3.hrbank3.dto.CursorPageResponseDto;
 import com.hrbank3.hrbank3.dto.audit_history.ChangeLogDetailDto;
 import com.hrbank3.hrbank3.dto.audit_history.ChangeLogDto;
+import com.hrbank3.hrbank3.dto.audit_history.ChangeLogSearchCondition;
 import com.hrbank3.hrbank3.dto.audit_history.DiffDto;
 import com.hrbank3.hrbank3.entity.EmployeeAuditHistory;
 import com.hrbank3.hrbank3.entity.enums.AuditSortField;
@@ -11,7 +12,6 @@ import com.hrbank3.hrbank3.entity.enums.AuditType;
 import com.hrbank3.hrbank3.event.EmployeeAuditEvent;
 import com.hrbank3.hrbank3.repository.EmployeeAuditHistoryRepository;
 import com.hrbank3.hrbank3.repository.EmployeeRepository;
-import com.hrbank3.hrbank3.repository.condition.ChangeLogSearchCondition;
 import com.hrbank3.hrbank3.repository.custom.EmployeeAuditHistoryRepositoryCustom;
 import java.time.Instant;
 import java.time.temporal.ChronoUnit;
