+++ src/main/java/com/hrbank3/hrbank3/service/EmployeeService.java
@@ -3,6 +3,7 @@ package com.hrbank3.hrbank3.service;
 import com.hrbank3.hrbank3.dto.employee.CursorPageResponseDto;
 import com.hrbank3.hrbank3.dto.employee.EmployeeCreateRequest;
 import com.hrbank3.hrbank3.dto.employee.EmployeeDto;
+import com.hrbank3.hrbank3.dto.employee.EmployeeSearchCondition;
 import com.hrbank3.hrbank3.dto.employee.EmployeeUpdateRequest;
 import com.hrbank3.hrbank3.entity.Department;
 import com.hrbank3.hrbank3.entity.Employee;
@@ -12,7 +13,6 @@ import com.hrbank3.hrbank3.event.EmployeeAuditEvent;
 import com.hrbank3.hrbank3.event.EmployeeNotificationEvent;
 import com.hrbank3.hrbank3.repository.DepartmentRepository;
 import com.hrbank3.hrbank3.repository.EmployeeRepository;
-import com.hrbank3.hrbank3.repository.condition.EmployeeSearchCondition;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
