+++ src/main/resources/static/index.html
@@ -1,14 +1,14 @@
 <!DOCTYPE html>
 <html lang="en">
-  <head>
-    <meta charset="UTF-8" />
-    <link rel="icon" type="image/ico" href="/logo.ico" />
-    <meta name="viewport" content="width=device-width, initial-scale=1.0" />
-    <title>HRBank</title>
-    <script type="module" crossorigin src="/assets/index-BL7rXp4c.js"></script>
-    <link rel="stylesheet" crossorigin href="/assets/index-53YfCFDQ.css">
-  </head>
-  <body>
-    <div id="root"></div>
-  </body>
+<head>
+  <meta charset="UTF-8"/>
+  <link rel="icon" type="image/ico" href="/logo.ico"/>
+  <meta name="viewport" content="width=device-width, initial-scale=1.0"/>
+  <title>HRBank</title>
+  <script type="module" crossorigin src="/assets/index-BL7rXp4c.js"></script>
+  <link rel="stylesheet" crossorigin href="/assets/index-53YfCFDQ.css">
+</head>
+<body>
+<div id="root"></div>
+</body>
 </html>
